requirements.txt
openai==0.27.2
pandas==1.5.2
python-dotenv==0.20.0